COZMO-16149 Code Lab Whats New off charger (#6698)
Fix bug where opening Code Lab directly from WhatsNew wouldn't have Cozmo drive off the charger - this was because the initial wake-up sometimes makes the robot wiggle off the contacts (e.g. during the severe-needs state on wakeup), so we cannot just check if on the charger at startup.

Track last world position and origin ID that robot was on charger
On opening Code Lab, drive clear of charger if Cozmo is within 10mm of the charger, not just if directly on the contacts.