VIC-863 Initial VisionScheduleMediatorComponent framework (#501)
Initial implementation of a new robot component which is meant
to mediate between the visionComponent and various systems which
may have need of various vision processing modes. Classes which
implement the IVisionModeSubscriber interface may send requests
for VisionModes to be enabled at Enumerated update frequencies
which correspond to integer updates every n frames, set in the
visionScheduleMediator_config.json file. The "Standard" update
setting is intended as a default setting for balanced performance
without having to specify High, Med, or Low.

I made a preliminary pass over the behaviors employed by
HighLevelAI as a test bed for the functionality. There is additional
work to be done to get the whole pallet of behaviors working via
this system, and to add Actions and BEIConditions as subscribers
to this system.