Move execution of `go get` to be external from cmake (#514)
* Move execution of `go get` to be external from cmake

`go get` examines source code and downloads external repositories
from Github or wherever that our source code depends on. When
running parallel builds on multiple Go targets at the same time,
they will both execute `go get` and could wind up (in the case of
our build servers, "could" is "frequently") cloning the same repos
into the same location on disk at the same time.

Without an easy way to govern the execution of that fetching
within CMake, this change moves the execution of `go get` to a
pre-build step running in a separate script. This first executes a
subset of `metabuild.py` to get a list of Go projects from BUILD.in
files* before sequentially running `go get` on each of them. Once
this is done (downloading takes about a minute but that would have
been incurred during the build previously anyway), subsequent runs
of `go get` take only a couple seconds, but can be skipped by
using the -I flag (ignore external dependencies) in future builds.

(*this is done because in a clean checkout, these lists don't exist
yet, but running `go get` should still be done before running the
full `metabuild.py` because it's important for that run to detect
the files we grabbed in `go get` so CMake can track them as
dependencies)