VIC-897 TEMP: Mega-Commit for OE Linux support
** DO NOT merge any branches containing this commit **

This is intended for devs to cherry-pick into their working
branches until we have upgraded all robots.

VIC-897 Add platform/cameraService for Camera access

CameraService manages an IPC connection to a server process running
on the robot to provide access to camera frames.

VIC-897 Replace AndroidHAL with CameraService

AndroidHAL has been dying a slow death. The only function it was
providing was camera access. It does not make sense to have this
component anymore. This commit removes androidHAL and replaces
it with cameraService, in the platform/camera directory. If we have
other libs & services that need to bridge between the OS and the
code in this repo, they should probably also live in the `platform`
dir.

Highlights:

- Use CameraService in VisionComponent to acquire frames

Make additional changes to support CameraService API:

- Frames acquired with CameraGetFrame must also be released.

To do this, we track the frame_id and pass it back to CameraService
when the frame is no longer being used.

TechDebt: Add missing array2d_impl.h for proceduralFaceTests

This allows us to compile the tests in Release.

VIC-897 Rename exes, update scripts for LE support

This is a minimal set of changes to build & deployment scripts
in order build code, deploy processes and run them on embedded
linux.

- Rename binary targets to vic-<name>
- Load platform config file in anim process
- Fix externals assets path for engine
- Install to /anki/{bin,lib,ect} + add env files for systemd
- Fix rsyncd to match /anki install target
- Remove unnecessary scripts
- Simplify aliases to just call systemctl (systemd control)
- Fix start-lldb script
- Update vscode debug configs
- Add compound vscode debugger targets to allow attaching to multiple
running processes at the same time.
- Update simpleperf scripts to work with the new binary names