VIC-897 Replace AndroidHAL with CameraService
AndroidHAL has been dying a slow death. The only function it was
providing was camera access. It does not make sense to have this
component anymore. This commit removes androidHAL and replaces
it with cameraService, in the platform/camera directory. If we have
other libs & services that need to bridge between the OS and the
code in this repo, they should probably also live in the `platform`
dir.

Highlights:

- Use CameraService in VisionComponent to acquire frames

Make additional changes to support CameraService API:

- Frames acquired with CameraGetFrame must also be released.

To do this, we track the frame_id and pass it back to CameraService
when the frame is no longer being used.