VIC-1122 Actions subscribe to VisionModes (#561)
These changes provide a framework for robot Actions to subscribe to
the new VisionScheduleMediator system for required VisionModes.
Actions are allowed to submit subscriptions in addition to behaviors
since Actions are considered to be independently functional. As such
each action should fully specify its required vision modes, if any, so
that it can be used outside the context of a behavior, or from within
the context of a behavior which does not require all the same
VisionModes as its subsequent Actions.

The tests in CST_BasicActions.cpp are reordered as part of these 
changes since the FaceObject test relies upon BlockWorld having 
discovered objects. If the default VisionModes are set to all off, as
they may be in the future, this would cause failures in the Basic 
Actions tests since there would be no identified blocks in BlockWorld
in the prior ordering of tests. The reordering ensures that victor has 
seen at least one block prior to running the FaceObject test.