VIC-1191 VisionScheduleMediator Debug Tools (#590)
These changes add a debug display to the Webots environment which
displays the currently active VisionModes from the
VisionScheduleMediator.

As part of adding the debug functionality, the UpdateVisionSchedule
function was reworked such that it first parses through the updated
data and then only changes the schedule and state of various
VisionModes as needed i.e. if a new subscription or release did not
change the UpdatePeriod or enabled state of any active modes, the
schedule is not rebuilt or re-sent to the VisionComponent.