Cleanup: combine "Behavior" and "BehaviorState" files (#574)
Several behaviors in the observing demo come paired as two files, a
"BehaviorName" version and a "BehaviorNameState" version. The latter
is usually a dispatcher that calls the former.
Here, any "BehaviorName" that appears self contained and not re-usable
is turned into an anonymous behavior "BehaviorNameInternal", defined
within "BehaviorNameState". Then, all "BehaviorNameState"s are renamed
as "BehaviorName". The result is one file for "BehaviorName" that might
delegate to an internally defined anon behavior "BehaviorNameInternal".