Fix double-closing of connections from server
When a new connection returned from a server is closed, the server
maintains a reference to the connection and will also try to close
it a second time (resulting in a panic) when the server is closed
and recursively closes its child connections. To fix this, we can
wrap the connections sent to clients through the `NewConns()`
channel in a helper type that will, before invoking the child
connection's `Close()` method, remove the child from the list of
connections the server is tracking.