VIC-334 Add Play state to observing (#565)
Adds a "Play" state to the observing system. The robot enters the state
when on the charger after a timer, or when off the charger when it sees
a cube and it hasn't played in a while. It performs a random play action
with the cube, or gives up, whichever comes first. It then returns to the
observing state.