VIC-1209 Handle VisionMode defaults from VSM (#603)
This change adds to the VisionScheduleMediator the ability to
subscribe to itself at init for VisionMode schedules that should be
our fallback when no outside subscriptions have been made. For now the
only internal subscription is DetectingMarkers at high rate. This
means there is no longer any active state for any VisionModes outside
the scope of the VSM (with limited and specific exceptions which access
the VisionComponent directly) and the vision_config.json defaults have 
been modified accordingly.