VIC-1150: Fix for small and short lift motions (#620)
Some small and short/quick lift motions are sometimes not rendered well off charger compared to on charger. This is because integral error is normally reset when the lift stops moving while on charger but not when it's off charger. When off charger, there's sometimes residual integral error from previous motions that needs to be unwound before the lift actually moves.

Integral error is now reset whenever a new height is commanded if the lift is not already in position. 

* Moved an encoder resolution check to SIMULATOR only since it accounts for so little of the commanded power anyway. (Still needed in sim because the lift oscillates and is forever considered to be moving.)

* Re-arranged some of the other integral error code to prevent unnecessary subtraction operation