Add script to automatically manage Go installations (#615)
* Add script to automatically manage Go installations

As @chapados recently discovered, our build breaks if run on Go
versions earlier than 1.9. To better manage building with Go, this
change adds a script to manage its installation in our ~/.anki dir
similar to what we do with CMake and the NDK.

To accomplish that, this change...
- extracts much of the functionality of cmake.py to a new
  toolget.py that can be used by any tool manager
- adds a corresponding go.py that downloads and installs a
  requested Go version. Unlike cmake.py, this tool stores a map of
  known SHAs for versions (Go doesn't appear to publish SHA txt
  files, or if they do the URL isn't immediately obvious), and will
  need to be updated at the same time we update our version
  requirements in the future.
- modifies build-victor.sh to export GOROOT (the environment
  variable that Go tools require to be set to the install location,
  if not the default `/usr/local/go`)
- modifies `run-go-get.sh` and `anki_build_go.cmake` to invoke
  `$GOROOT/bin/go` in place of assuming that `go` would be in the
  `$PATH`.

One potential downside is that the last point takes the ability to
custom-specify a Go version/installation out of the hands of the
person building. Not sure if there's a great solution for this,
since an older Go version installed via `brew` or whatever (which
is more likely to be in someone's existing `$PATH`/`$GOROOT` setup)
_should_ be ignored in favor of our Anki-installed version in most
cases.

* Add option to specify a custom GOROOT

If the `$GOROOT` environment variable is set going into the build
script, that Go installation will be used instead of the Anki-
managed installation. A warning will be printed if the custom
version is not equal to the default version, which can be safely
ignored if the change is intentional.