Trigger multiple behaviors on shake (#619)
Currently, if you shake the robot, it will enter its HighLevelAI
behavior. We have or will soon have some other behaviors that are
ready for testing, but not ready to be situated in their final
resting spot in the behavior system.
This changes the DevDispatchAfterShake behavior to trigger the n-th
behavior in its list after shaking the robot n times. Alternatively,
the console var DevDispatchAfterShake will trigger the same index.
ReactToSound is now triggered after two shakes.
The alternative is to send the desired behaviorName through webots,
but that requires connecting and all that.