VIC-1316: Use default schedules for unspecified modes in VisionScheduleMediator (#648)
Toggling vision modes with console vars (and webots messages) was broken because even if the mode was toggled "on", the VisionScheduleMediator was using "false" for all the underlying schedules. So modes were "enabled but unscheduled". This just uses default schedules when modes are unspecified, which leaves those modes scheduled and allows enabling them to work again.