VIC-1293 Faster, simpler deployment, now in one script.
Our deployment process consisted of 2 scripts and used multiple
rsync commands with a set of options that results in inefficient transfers.

These changes modify the build system to output files into directories
that match structure expected under `/anki` on robots. This in turn
makes it possible to build a filelist that is then passed to a single
rsync command, avoiding the cost of multiple process + network connection
startup and teardown. This change also removes the old, complicated logic
used to support having multiple deployment directories on the robot.
Given the use of rsync, we now only support a single assets dir
(in /anki/data/assets) that will be synchronized with the contents in
the appropriate build folder (controlled by the -c option to deploy.sh)

Finally, update rsync options: -rlptD -uzvP

These options ensure that timestamps, links, permissions
and special files are preserved. In addition, we now use compression
(-z) and most importantly -u, which causes rsync to skip files that
are newer on the receiver.

Combined together, these changes results result in an approximate speedup
of 25X for small-change deployments and slightly faster performance for
full installs (due to -z flag).