VIC-1218 Fix WaitForImagesAction invalid subscription to VSM (#664)
The WaitForImagesAction may be invoked with VisionMode::Count to
indicate that the user wants to wait for N images to be processed
irrespective of which vision processing modes are enabled. The VSM is
not equipped to handle requests for VisionMode::Count, so
WaitForImagesAction should simply not subscribe to anything under
these circumstances.