Cloud process refactor (#671)
* Fix double-closing of connections from server

When a new connection returned from a server is closed, the server
maintains a reference to the connection and will also try to close
it a second time (resulting in a panic) when the server is closed
and recursively closes its child connections. To fix this, we can
wrap the connections sent to clients through the `NewConns()`
channel in a helper type that will, before invoking the child
connection's `Close()` method, remove the child from the list of
connections the server is tracking.

* Update Chipper API, Client -> Stream

- Update voicetest program for new API
- Add a couple voice samples for "do a fist bump" and "set a timer"
  intents
- Update Go dependency versions

* Refactor cloud process inputs/outputs

The previous cloud process API assumed too many details about
communication implementation, by taking in IPC connections to
receive audio input and send cloud responses. It also lacked the
ability to receive data from a variable number of sources, when we
may want to have test channel inputs alongside the microphone
input.

This change:
- splits the voiceContext code into a new file for better
  organization
- adds send.go and receive.go which define new APIs for sending and
  receiving data to/from the cloud process
- adds the "-test" flag to the main cloud process executable to
  start it with a test channel enabled (on by default, should be
  disabled for shipping, however we plan to do that)
- changes existing programs that use the cloud process to use the
  new API (instantiating a cloud.Process{} and then adding data
  sources to it, instead of passing things into RunProcess()
  function)
- prepares code to send intents as json structures instead of
  simple strings (not yet enabled, needs corresponding AI changes
  before this can be completed)