Add ability for godeps tool to remove unused dependencies (#692)
I noticed that with latest code, several dependencies in godeps.json
are actually no longer used by our Go code. This updates the godeps
tool so that when it detects this situation, it will automatically
remove the unused dependencies from being tracked and halt the build
the same way it does when adding a new dependency.