VIC-1144 Add BEIConditions to VSM subscription Model (#662)
These changes provide a framework for IBEIConditions to manage
functionality that should only be active at certain times. A key use
case for this feature is subscription to various VisionModes through
the VisionSchedule mediator. These transitions are managed at the
level of IBEICondition through a new SetActive interface. Conditions
must be SetActive prior to evaluation, and should be set inactive any
time they are not in scope for evaluation.