Fix unique behavior debug labels (#701)
If MakeParamTunable is called during the constructor, the debug label
hasn't yet been made unique, so only the first instance's var is
registered.
This returns to the implementation with a static map. An alternative
is passing the debugLabel to the ctor, but that's a huge change for a
small purpose. Anotion option is to stick it in the json, but it's a
const reference, and we don't want a bunch of copies.