CLAD: Adds option for unions to allow non-distinct types (#660)
Adds a keyword for CLAD unions to specify that they may contain
non-distinct member types.
Starting with an example,
```
union MyUnion {
    StructureA s1, // note: the same type
    StructureA s2, // note: the same type
    StructureB s3,
}
```
then as usual, `MyUnion` can store data, based on the type in
`enum MyUnionTag`, which takes on values `s1`, `s2`, or `s3`.
This works fine with default CLAD arguments, but doesn't work
with the flag `--output-union-helper-constructors`. This flag
generates constructors like `MyUnion( StructureB&& );` These
ctors set both the data fields and the union tag. For unions
with non-distinct member types, as in the above example, the
ctor `MyUnion( StructureA&& )` would appear twice. Even if it
appeared once, it would not be able to specify the `MyUnionTag`
tag, so it would effectively be useless. These ctors are useful
when the union member types are known to be distinct, but it would
also be useful to allow non-distinct member types in some unions,
without globally disabling `--output-union-helper-constructors`.
In this commit, unions can explicitly allow for non-distinct member
types, even when `--output-union-helper-constructors` is passed,
using the following syntax:
```
union dupes_allowed MyUnion {
    StructureA s1, // note: the same type
    StructureA s2, // note: the same type
    StructureB s3,
}
```
(This has no effect if --output-union-helper-constructors is omitted.)
In this case, assuming --output-union-helper-constructors is used,
the ctors `MyUnion( StructureA&& )` are selectively omitted, but the ctor
`MyUnion( StructureB&& )` is still generated.