COZMO-15658: Crash in CozmoButton::ShowEnabledState (#6758)
Only 1 thing I can see that could be causing a NullReferenceException in this method, although I'm not sure how gameObject could
be null. I'm guessing it is because I'm not finding any of the error logging in DAS, but we're still getting several crashes per
day from this code.

Open to suggestions on what else could be causing a NRE here.