Temporary app to engine webservice channel (#716)
Meet victor is the first feature for VS where we'll need the engine and
app to exchange data. That feature has had its kickoff, but we're still
waiting for the connection flow and auth/non-auth channels betwen the
app and engine to be complete.
Instead of pausing Meet Victor development, this adds a **temporary**
pathway for the app and engine to pass data. The app can GET from
http://robot:8888/sendAppMessage?type=CladTag&field1=val1&field2=val2
to pass a message to the engine. (This URL must be valid, e.g., spaces
replaced with + and other characters escaped.)
The app can periodically request a list of messages that the engine has
sent it with
http://robot:8888/getAppMessages
which returns a json array of serialized clad structs, one for each
message.
On the engine side, the new AppToEngineHandler translates all of that
into MessageGameToEngine's or MessageEngineToGame's, and broadcasts it
or handles it, respectively. This should make the swap to the real
app-engine channel easier, especially if we keep using
MessageGameToEngine and MessageEngineToGame for app messages.
Note: since MessageEngineToGame and MessageGameToEngine contain
messages instead of structures, we have to manually (de)serialize
them, which sucks.