VIC-685 (observing voice commands) Unclaimed intents (#726)
If a user intent goes unclaimed, it means we haven't yet added a
transition to handle it.
It would be useful at this stage if the robot performs a reaction, so
we can catch these missing transitions early. This adds a BEICondition
for unclaimed intents.
This also means that we could, in theory, skip creating transient states
in HL-AI to handle intents that don't make sense in the context. For
instance, if the robot is on its charger, and you say go to your
charger, it should go "huh?". That could either be a transition triggered
by a "go_to_charger" intent from OnCharger to OnChargerConfused, and
eventually back to OnCharger. OR it could just be an unclaimed intent
handled higher up. The former is better, but would clutter up the state
transitions if added at this stage. So for now, the context-aware "huh?"
is the same as "intent unclaimed."