C++11 console funcs
Adds the option for an std::function instead of a function pointer for
IConsoleFunction, which is useful if you have a capture list.
From reading and testing, the std::function takes up the same size as
the function pointer, even if you capture big variables (that goes on
heap). Plus, nothing should change in release builds.