Behavior App Intents (#744)
Creates a MessageGameToEngine called AppIntent for the app to request
that the robot perform something. These are currently being passed
through the webservice as a querystring, so this keeps the AppIntent
simple for now, with one param "param" for the "intent" of that name.
Once the real channel is in place through the Switchboard, AppIntent
can be a union of strucs.
The UserIntentComponent now listens for the AppIntent message and sets
the corresponding user intent as pending, based on a json-defined
mapping (copied from the cloud->user intent mapping). The sanitization
step is also copied, which is (temporarily) exploited here so that the
generic "param" of the AppIntent can be mapped to the corresponding
field of the UserIntent. This won't be needed with a union of structs.
Also updates webviz for app intents.