Add baked-in certificates to cloud binary (#764)
GRPC connections are failing on DVT2/LE robots due to the system
not having root certificates available for Go to use for TLS.
@gwatts helpfully supplied a package he wrote that bakes certs into
the binary, which allows the robot to connect to the cloud again.

I initially implemented this in a Linux-only file, but then
discovered that our Go dependency checks assume they're running on
Mac, and would have to add successive calls with GOOS="android" to
some command line calls to do full checking on both OSes. That work
added ~4 seconds (roughly double the current time) to Go dependency
checking. As such, I decided to add the certificate stuff on both
platforms, but left the extra script work checked in (and commented
out) so it's still there if we need to use it in the future.

Because we no longer have control (at a source control level) over
process startup, this change also makes some debug-ish flags on the
cloud process (opening a test channel for programs like wavtest to
use to connect to, and doing verbose logging) the default behavior.