Fix victor not staying on charger (#774)
After returning to his charger, Victor momentarily pauses and then
drives off. Then it tries again, acting like a cat unsatisfied with
its litter box. This is because the behavior consists of two nested
dispatchers, each of which will always want to run. So even though
the robot is already on its charger, the bottom dispatcher says it
wants to run, and the behavior repeats instead of exiting to
ObservingOnCharger.
The fix adds a condition in the lower dispatcher and to one of the
behavior classes to not be activated if already on the charger. Now
the behavior exits to ObservingOnCharger.