Shared named behavior timers (#781)
Currently all cooldowns are local to some condition or behavior. If we
want to have two similar behavior instances, one for voice commands and
one for robot-instagated actions, currently they would have to have
different cooldowns. For instance, "Cozmo, roll that block" might
trigger one behavior, but another, perhaps more timid, roll block
behavior could be triggered if a cube is discovered during
exploring/hiking. If the voice command was just issued for roll cube,
we probably don't want the other roll cube behavior to be chosen.
This adds a behavior component for shared timers, a condition to check
the elapsed time of a timer, and a way to reset timers from a behavior.
Fist bump and PlayingWithCube now use these.