VIC-455 rm stale assets as last step of asset build (#771)
Our asset build currently consists of just copying assets from
various locations in the source tree to a `data/assets` folder
under the build output dir.  In the case where an asset is removed
from the source tree between builds, there is no logic in
CMake to cause the deleted asset to be removed from the
build folder.  To address this issue, we previously force-deleted
all assets in the build folder when either `-f` or `-X` was
specified.  Using these options was therefore time consuming.

This approach uses CMake to address the issue as follows:

- Keep a cumulative list of all output asset files during the
  cmake configure phase.

- Write the expected output file list to an asset-build.manifest
  file using the `configure_file` method. This also happens
  during the CMake configure phase, *not* during the build phase.

- Adds a custom command and target that depends on the dst files
  and a "stamp" file indicating that the stale assets have been
  deleted.  The custom command calls a shell script that compares
  the manifest with the dir contents, deletes files not in the list
  and write the date + names of deleted files into the .stamp file.

Caveats:

- The file of know destination files is generated during the configure
  phase, since this is the only time that we are guaranteed to process
  the entire list.  However, since all of the assets in our project are
  specified using rules in BUILD.in files, it will typically be necessary
  to specify the `-f` flag to cause CMake to re-run the configure step
  and update the asset-build.manifest.

related to VIC-456