Unit test HighLevelAI voice commands (#803)
Tests that every cloud and app intent that we consider "done" has
been properly hooked up in BehaviorHighLevelAI. For each cloud
or app intent, it searches for a UserIntentPending with the
corresponding UserIntentTag in all transitions out of every state in
BehaviorHighLevelAI. If a matching transition is found, that user
intent tag must appear in an explicit list we consider "done." If
a tag that we consider "done" does not have a transition out of a
state, the test fails, unless that state was explicitly listed as an
exception (for, e.g., intentional context-dependent "huh?" reactions).
A friend declaration is used to peek into the user intent pending
condition. But since we also want to peek into a nested class of
InternalStatesBehavior, whose definition is in a .cpp, this introduces
a UnitTestKey, which is an object that only unit tests can create.
If an engine class exposes a dev method that takes this as a parameter,
we can be sure it is only used in unit tests. This is used to list
all the conditions in InternalStatesBehavior, but hopefully would be
useful elsewhere.