VIC-1322 Fix clad vs clean builds (#796)
* VIC-1322 Fix clad vs clean builds

The work for cleaning clad on OverDrive was two part:
1) only write generated files (.cpp, .h, .cs) when the contents were different
2) delete unused generated files

The first step would prevent re-compilation for no-change, i.e. regenerating
the exact same file over the original file dirties the time-stamp and
forces the compiler to recompile any file that includes it.

OverDrive had this problem as all clad files were generated in bulk,
Victor does not have this issue as each individual clad file is generated
separately and only if the time-stamp is dirty.

The second step applies. Any clad files that are removed from the source
list will have orphan generated files, left over .cpp, .h, .cs that are not
generated from current source. If these files are referenced compilation
errors will occur on a clean build or build that never had the original files.

Solved for Victor by globbing the generated directory for clad files and
then checking all targets and all source files of targets for a match.
If a generated clad file is no-longer in the source list for any target
it is orphaned and should be deleted. This process can occur during
cmake configuraton.

* PR review