COZMO-16367 Code Lab vertical: Horizontal XML project blocks disappear (#6781)
With the Feb 2018 Scratch merge, the Scratch team has encouraged us to use new code for the vm workspaceUpdate event. One byproduct of the new code is that projects serialized as XML can be intermittently removed from the screen. I think it is a race condition between getting the vm workspaceUpdate event for the initial empty project, and getting it for the desired project but I haven't 100% proven it. In any case, using the old code instead for XML projects seems to avoid this bug.