COZMO-16510 Load CL serialized featured project into memory only when needed (#6791)
When pretty-printing came in (hurrah!), the code was also changed to load the featured content serialized json data in memory throughout the life of CodeLabGame. Now that we are running into memory pressure (causing crashes in some cases), we want to reduce what is in memory.

We know that having too many user projects in memory causes crashes. Since the average user doesn't have more than 1 user project, it makes more sense to start by removing featured projects from memory and only reading them from disk when needed, as it was previously.

Each json file in featuredProjects is pretty-printed json. When calling configure.py, each file will be copied to encodedFeaturedProjects with the ProjectJSON field set to non-pretty-printed serialized json.