Rework Go dependency pulling to be faster, more reliable (#853)
* Clone dependency repos to desired commit before using go get

If a go package we depend on is deleted in a newer commit than the
commit we've versioned, then `go get` will fail to find that
package (and error out) because it always pulls the newest commit
when doing a fresh clone.

To fix this, we will do the cloning of dependencies ourselves
before running `go get`.

* Update dependency script to run `go get` itself

Rather than run the dependency script to clone repositories, use
another script to run the `go get` command, and then run the
dependency script again to verify that repos are versioned, it
makes more sense to simply do all of these steps within the
dependency script.

This combining reduces the typical case build script time for Go
verification ("typical case" being where you already have the
dependencies you need at the right commit) from 4.1 seconds, using
the old multiple-scripts approach, to 1.7 seconds.

* Use one global GOPATH for all of Victor

The initial Go build setup was written in a way such that different
Go applications could use different GOPATH directories. However, in
practice, I've yet to set a different GOPATH and yet to see a use
case for even doing so, and as we add/modify Go tools it seems
silly to keep supporting this potential use case. So, this commit
makes it so that `cloud/go` is now the global GOPATH for all Victor
projects.

Obviously, if we ever use Go in Victor outside of cloud stuff, we
can move it to a less cloud-specific location.

* Quit on clone error