VIC-1430 Six easy json voice commands + Waking up (#839)
VIC-1430 Six easy json voice commands + Waking up

+ Some voice commands in json: Roll Cube, Go to Sleep, Do a trick,
  Hello, Good Morning, Goodbye, Goodnight.
+ If the robot is napping, "Hey Cozmo" should wake it up, instead of
  having to say "wake up." This adds a special case to
  BehaviorCoordinateGlobalInterruptions to ignore the regular wake word
  behavior if the Napping state is active in HLAI. Instead, the a new
  transition out of Napping occurs when the trigger word is pending, and
  the regular BehaviorUpdate() loop of HLAI clears the pending trigger.
+ Also adds some state replacement rules for when re-activating HLAI,
  so it won't immediately return to napping, for instance.
+ For roll cube, modifies LookForFaceAndCube to optionally break if
  there is a cube in blockworld, so that the robot will look for a
  cube only if one is not known, or only until one is found.
+ The play_specific test has two flavors now in HLAI, and the
BehaviorHighLevelAI.VoiceIntentsHandled test needs to test both. The
test now compares a sample list of cloud intents to both the intents
in user_intent_map and those intents that are hooked up in HLAI. If
any intents are hooked up but not listed, it fails. If any tests are
not hooked up but appear in the completed list, it fails unless the
HLAI state is listed as an exception. For app intents, we look for
discrepancies only between app intents in user_intent_map and the
hooked up commands in HLAI, using the cloud intent samples as a
guide.
Also fixes the problems with HLAI that the test found.