VIC-1593: victor-ble-cli - Handle robots named Vector (#890)
We intend to name the robot Vector for production use.  If the tool
sees a robot advertising the name "Vector XYXY", treat it the same as
those advertising the name Victor.

Also added support for quoting arguments on the command line. This
allows a command like:

connect "Vector R2D2"

to be parsed as having 2 arguments instead of 3.  Should also
help for setting the wifi configuration when the ssid or password has
a space in it.