Temporarily disable compression (#892)
The use of compression is causing an intermittent issue when using
Google as our speech provider in which it claims we're streaming
audio faster than real-time, which we're not. There are other
reports of this issue on the internet that lead us to believe it
might be a problem on Google's end when dealing with compressed
audio. Until we either resolve this or decide to switch to MS as
our default voice provider, we should disable compression.