VIC-1634 vic-ble-cli - support 16-bit service UUID
I am updating the advertising code on Victor to use one
of anki's 16-bit service UUIDs.  This allows us to cram
all the advertising information into 1 packet and avoid
relying on the scan response.

This update allows the vic-ble-cli tool to find both old
and new robots advertising and offering either the old
128-bit service UUID or the new 16-bit service UUID.