Mwillett/vic 1257 support generic point set inserts to nav map (#893)
This provides a generic interface for geometric objects commonly used to to test spacial constraints in Euclidean space. Additionally, inheritance from these base classes will allow for greater flexibility when handling common geometric objects without having to build out additional boilerplate code in methods that may want to use similar types.

First, we provide a abstract base class PointSet that can be used to test any generic collection of points in n-dimensional space, both continuous or discreet.

This is extended by a second abstract class for ConvexPointSets, which are useful for testing linear constraints of the form Ax+b > 0

We then provide a couple of new instances for ConvexPointSets, as well as extend existing LineSegment, FastPolygon, and AxisAlignedQuad classes using these new types.

In the future, classes such as Rect, RotatedRect, Triangle, and ConvexPolygon could all be implemented as instances of ConvexPointSet for allow for polymorphic operations on these classes.