COZMO-16527 Fix Code Lab projects intermittently not running (#6814)
Alter how we put the green flag on the workspace in order to try to fix:
COZMO-16527 - [Code Lab Horizontal] - New project - Sometimes Cozmo does nothing on running a created script and script doesn't show after save/load
COZMO-16178 - [Code Lab Vertical] - Workspace - The default "when Green Flag clicked" block disappears on quickly entering a new project

For COZMO-16527, when the bug occurred, the green flag on the workspace would not be recognized by the vm. I think the problem is that the green flag was put on the workspace too early during workspace construction.

In this change, ensureGreenFlagIsOnWorkspace is called at most one time. It used to be called a lot more as it was called when blocks were deleted, even during project load time when a project could be wiped from the workspace.

ensureGreenFlagIsOnWorkspace is now called on a significant delay (1 second) so Scratch can finish setting up the workspace. This is a hack but I haven't found another solution yet.

In this change, we no longer put the green flag back on the workspace after it is deleted in horizontal. We can revisit this change later if desired.